create utils folder and move things there
