+++ baby-gru/src/components/BabyGruContainer.js
@@ -3,7 +3,7 @@ import { Navbar, Container, Nav, Tabs, Tab, Accordion, Button, Col, Row, Card, S
 import { BabyGruDisplayObjects } from './BabyGruDisplayObjects';
 import { BabyGruWebMG } from './BabyGruWebMG';
 import { v4 as uuidv4 } from 'uuid';
-import { BabyGruCommandCentre, convertRemToPx, convertViewtoPx } from '../BabyGruUtils';
+import { BabyGruCommandCentre, convertRemToPx, convertViewtoPx } from '../utils/BabyGruUtils';
 import { BabyGruButtonBar } from './BabyGruButtonBar';
 import { BabyGruFileMenu } from './BabyGruFileMenu';
 import { BabyGruRamachandran } from './BabyGruRamachandran';
