+++ baby-gru/src/components/BabyGruFileMenu.js
@@ -1,8 +1,8 @@
 import { NavDropdown, Form, Button, InputGroup, Modal, FormSelect, Col, Row, Overlay, Card, FormCheck } from "react-bootstrap";
 import { BabyGruMolecule } from "./BabyGruMolecule";
-import { BabyGruMap } from "./BabyGruMap";
+import { BabyGruMap } from "../utils/BabyGruMap";
 import { useEffect, useState, useRef, createRef } from "react";
-import { BabyGruMtzWrapper, cootCommand, readTextFile } from '../BabyGruUtils';
+import { BabyGruMtzWrapper, cootCommand, readTextFile } from '../utils/BabyGruUtils';
 import { InsertDriveFile } from "@mui/icons-material";
 import { BabyGruMoleculeSelect } from "./BabyGruMoleculeSelect";
 import { BabyGruImportDictionaryMenuItem, BabyGruImportMapCoefficientsMenuItem, BabyGruDeleteEverythingMenuItem, BabyGruLoadTutorialDataMenuItem } from "./BabyGruMenuItem";
