+++ baby-gru/src/components/BabyGruMenuItem.js
@@ -2,8 +2,8 @@ import { MenuItem } from "@mui/material";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { OverlayTrigger, Popover, PopoverBody, PopoverHeader, Form, InputGroup, Button, FormSelect, Row, Col } from "react-bootstrap";
 import { SketchPicker } from "react-color";
-import { BabyGruMtzWrapper, readTextFile } from "../BabyGruUtils";
-import { BabyGruMap } from "./BabyGruMap";
+import { BabyGruMtzWrapper, readTextFile } from "../utils/BabyGruUtils";
+import { BabyGruMap } from "../utils/BabyGruMap";
 import { BabyGruMolecule } from "./BabyGruMolecule";
 import { BabyGruMoleculeSelect } from "./BabyGruMoleculeSelect";
 import BabyGruSlider from "./BabyGruSlider";
