+++ baby-gru/src/utils/BabyGruMap.js
@@ -1,4 +1,4 @@
-import { cootCommand, postCootMessage, readDataFile } from "../BabyGruUtils"
+import { cootCommand, postCootMessage, readDataFile } from "./BabyGruUtils"
 import { readMapFromArrayBuffer, mapToMapGrid } from '../WebGL/mgWebGLReadMap';
 
 export function BabyGruMap(commandCentre) {
